(feat) Skip serialization of `None` values
This omits fields from JSON payloads when the value `is_none()`. For
the read-only values given to the square api it will sometimes silently
ignore the value and other times it will outright fail because you are
not supposed to be supplying a value for a square-managed field (i.e.
the `version` field).

Some fields _must_ be ommitted from the payload based on the value of
other fields. For example, you cannot send a `next_payment_amount_money`
if you have 1 configured `payment_request` for `BALANCE`... but if you
send `next_payment_amount_money` you have to have a `Money` value...
`null` does not work.

This does not prevent any parsing on _deserialization_, though. The
values should always be present in the resulting struct as they have
been from the beginning.